Reorganize query backend files
- move code for building index to its own directory
- move testing code to its own directory
- simplify the backend_domain_* files
