+++ backend/domain_config/__init__.py
@@ -1,16 +1,12 @@
 import os
 DOMAIN = os.environ.get('LENSING_DOMAIN', 'wikipediahistory')
 
-import avherald.backend_domain_config
-import avherald.backend_domain_settings_defaults
-
-import wikipediahistory.backend_domain_config
-import wikipediahistory.backend_domain_settings_defaults
+import avherald.domain_config
+import wikipediahistory.domain_config
 
 if DOMAIN == 'avherald':
     _import_from = avherald
 else:
     _import_from = wikipediahistory
 
-backend_domain_config = _import_from.backend_domain_config
-backend_domain_settings_defaults = _import_from.backend_domain_settings_defaults
+domain_config = _import_from.domain_config
