--- backend/domain_config/avherald/backend_domain_settings_defaults.py
@@ -1,12 +1,0 @@
-"""
-Domain-specific adjustments to the default server settings.
-"""
-
-import avherald
-
-settings = {
-  'querier': {
-    'fields_to_prime': ['role', 'predicate', 'event', 'year'] + avherald.facet_field_names,
-    'fields_for_text_searches': ['description', 'role', 'predicate', 'event', 'year'] + avherald.facet_field_names
-  }
-}
