+++ backend/src/backend
@@ -32,7 +32,7 @@ except ImportError:
 # import backend_settings
 # import backend_settings_defaults
 # handles which domain to choose based on environment variables internally
-from domain_config import backend_domain_settings_defaults
+from domain_config import domain_config
 
 logging.basicConfig(stream=sys.stdout,
                     level=logging.DEBUG)
@@ -67,14 +67,14 @@ def create_app(index, redis_address, redis_port):
     cache = RedisCache(redis_address, port=redis_port, default_timeout=0,
                        key_prefix="query_cache")
     querier = queries.Querier(whoosh_index, cache,
-                              **backend_domain_settings_defaults.settings.get('querier', {})) # noqa
+                              **domain_config.settings.get('querier', {})) # noqa
 
     querier.prime()
 
     @QueryRequest.application
     def application(request):
         querier = queries.Querier(whoosh_index, cache,
-                                  **backend_domain_settings_defaults.settings.get('querier', {})) # noqa
+                                  **domain_config.settings.get('querier', {})) # noqa
 
         try:
             query = request.json
