--- backend/src/backend_settings.py
@@ -1,53 +1,0 @@
-"""
-Utilities for reading and using settings.
-"""
-
-import sys
-import backend_settings_defaults
-
-class UnknownSetting(Exception):
-  def __init__(self, setting):
-    Exception.__init__(self, "unknown setting \"%s\"" % (setting))
-
-class MissingRequiredSetting(Exception):
-  def __init__(self, setting):
-    Exception.__init__(self, "missing required setting \"%s\"" % (setting))
-
-def read_from_file(path):
-  """
-  Reads in settings from a text file.
-  """
-  def check_settings(settings, defaults):
-    for setting, value in settings.iteritems():
-      if setting not in defaults:
-        raise UnknownSetting(setting)
-      def_value = defaults[setting]
-      if hasattr(def_value, 'iteritems'):
-        check_settings(value, def_value)
-  with open(path) as input_file:
-    settings = eval(input_file.read())
-    check_settings(settings, backend_settings_defaults.settings)
-    return settings
-
-def apply(dest, settings, *defaults):
-  """
-  Apply settings as properties of an object, falling back on defaults as needed.
-  """
-  for defaults_set in defaults:
-    for setting, value in defaults_set.iteritems():
-      if setting not in settings:
-        setattr(dest, setting, value)
-  for setting, value in settings.iteritems():
-    if any(setting in ds for ds in defaults):
-      setattr(dest, setting, value)
-    else:
-      raise UnknownSetting(setting)
-
-"""
-Make a class to expand settings to an object with given defaults filled in.
-"""
-def make_settings_structure(*defaults):
-  class SettingsStructure:
-    def __init__(self, **settings):
-      apply(self, settings, *defaults)
-  return SettingsStructure
